Moving CNVError into seabird.exceptions
